requirements.txt
# Core Dependencies
numpy==1.24.3
scipy==1.11.1
pandas==2.0.3

# Signal Processing
mne==1.4.2
pywavelets==1.4.1
scipy-signal==0.1.0

# Machine Learning
scikit-learn==1.3.0
joblib==1.3.1
imbalanced-learn==0.11.0

# Deep Learning (Optional for advanced models)
torch==2.0.1
tensorflow==2.13.0

# Web Framework & API
fastapi==0.100.0
uvicorn[standard]==0.23.1
websockets==11.0.3
python-multipart==0.0.6
pydantic==2.0.3

# Database
psycopg2-binary==2.9.6
sqlalchemy==2.0.19
alembic==1.11.1

# Redis for caching
redis==4.6.0
hiredis==2.2.3

# Time Series Database
timescaledb==0.1.0

# Data Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# Real-time Processing
pyzmq==25.1.0
asyncio==3.4.3

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0
toml==0.10.2

# Logging & Monitoring
loguru==0.7.0
prometheus-client==0.17.1

# Testing
pytest==7.4.0
pytest-cov==4.1.0
pytest-asyncio==0.21.0
hypothesis==6.82.0

# Code Quality
black==23.7.0
flake8==6.0.0
mypy==1.4.1
pylint==2.17.4

# Documentation
sphinx==7.0.1
sphinx-rtd-theme==1.2.2

# Utilities
tqdm==4.65.0
click==8.1.6
python-dateutil==2.8.2
pytz==2023.3

# Security
cryptography==41.0.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Email & Notifications
sendgrid==6.10.0
twilio==8.5.0

# File Handling
h5py==3.9.0
pyedflib==0.1.30

# API Documentation
fastapi-swagger-ui==0.2.0

# Performance
numba==0.57.1

# UI
streamlit>=1.37
PyPDF2>=3.0.0
